+++ lib/ui/auth/auth_complete_sheet.dart
@@ -32,7 +32,7 @@ class _AuthCompleteSheetState extends State<AuthCompleteSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),