+++ lib/ui/auth/auth_confirm_sheet.dart
@@ -110,7 +110,7 @@ class _AuthConfirmSheetState extends State<AuthConfirmSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),