+++ lib/ui/generate/generate_confirm_sheet.dart
@@ -89,7 +89,7 @@ class _GenerateConfirmSheetState extends State<GenerateConfirmSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),