+++ lib/ui/handoff/handoff_confirm_sheet.dart
@@ -108,7 +108,7 @@ class _HandoffConfirmSheetState extends State<HandoffConfirmSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),