+++ lib/ui/register/register_complete_sheet.dart
@@ -39,7 +39,7 @@ class _RequestCompleteSheetState extends State<RequestCompleteSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),