+++ lib/ui/send/send_complete_sheet.dart
@@ -40,7 +40,7 @@ class _SendCompleteSheetState extends State<SendCompleteSheet> {
               height: 5,
               width: MediaQuery.of(context).size.width * 0.15,
               decoration: BoxDecoration(
-                color: StateContainer.of(context).curTheme.text10,
+                color: StateContainer.of(context).curTheme.text20,
                 borderRadius: BorderRadius.circular(5.0),
               ),
             ),