+++ lib/ui/send/send_sheet.dart
@@ -493,7 +493,7 @@ class _SendSheetState extends State<SendSheet> {
                       height: 5,
                       width: MediaQuery.of(context).size.width * 0.15,
                       decoration: BoxDecoration(
-                        color: StateContainer.of(context).curTheme.text10,
+                        color: StateContainer.of(context).curTheme.text20,
                         borderRadius: BorderRadius.circular(5.0),
                       ),
                     ),