+++ lib/ui/settings/changerepresentative_sheet.dart
@@ -324,7 +324,7 @@ class AppChangeRepresentativeSheet {
                                     height: 5,
                                     width: MediaQuery.of(context).size.width * 0.15,
                                     decoration: BoxDecoration(
-                                      color: StateContainer.of(context).curTheme.text10,
+                                      color: StateContainer.of(context).curTheme.text20,
                                       borderRadius: BorderRadius.circular(5.0),
                                     ),
                                   ),