+++ lib/ui/settings/changerepresentativemanualentry_sheet.dart
@@ -122,7 +122,7 @@ class _ChangeRepManualSheetState extends State<ChangeRepManualSheet> {
                             height: 5,
                             width: MediaQuery.of(context).size.width * 0.15,
                             decoration: BoxDecoration(
-                              color: StateContainer.of(context).curTheme.text10,
+                              color: StateContainer.of(context).curTheme.text20,
                               borderRadius: BorderRadius.circular(5.0),
                             ),
                           ),