+++ lib/ui/transfer/transfer_overview_sheet.dart
@@ -94,7 +94,7 @@ class AppTransferOverviewSheet {
                                   height: 5,
                                   width: MediaQuery.of(context).size.width * 0.15,
                                   decoration: BoxDecoration(
-                                    color: StateContainer.of(context).curTheme.text10,
+                                    color: StateContainer.of(context).curTheme.text20,
                                     borderRadius: BorderRadius.circular(5.0),
                                   ),
                                 ),