+++ lib/ui/widgets/remote_message_sheet.dart
@@ -46,7 +46,7 @@ class _RemoteMessageSheetStateState extends State<RemoteMessageSheet> {
                       height: 5,
                       width: MediaQuery.of(context).size.width * 0.15,
                       decoration: BoxDecoration(
-                        color: StateContainer.of(context).curTheme.text10,
+                        color: StateContainer.of(context).curTheme.text20,
                         borderRadius: BorderRadius.circular(5.0),
                       ),
                     ),